improvements

+++ src/interfaces/IMidnight.sol
@@ -47,7 +47,7 @@ struct ObligationState {
     uint256 withdrawable;
     bool created;
     uint16[7] fees;
-    uint64 continuousFee;
+    uint32 continuousFee;
 }
 
 struct Position {
